+++ v0/destinations/facebook_pixel/transform.js
@@ -26,67 +26,14 @@ const {
   getValidDynamicFormConfig,
   isDefinedAndNotNull
 } = require("../../util");
-const { isDefined } = require("rudder-transformer-cdk/build/utils");
 
-/**  format revenue according to fb standards with max two decimal places.
- * @param revenue
- * @return number
- */
+const {
+  deduceFbcParam,
+  formatRevenue,
+  getContentType,
+  transformedPayloadData
+} = require("./utils");
 
-const formatRevenue = revenue => {
-  const formattedRevenue = parseFloat(parseFloat(revenue || 0).toFixed(2));
-  if (!isNaN(formattedRevenue)) {
-    return formattedRevenue;
-  }
-  throw new CustomError("Revenue could not be converted to number", 400);
-};
-/**
- *
- * @param {*} message Rudder Payload
- * @param {*} defaultValue product / product_group
- * @param {*} categoryToContent [ { from: 'clothing', to: 'product' } ]
- *
- * We will be mapping properties.category to user provided content else taking the default value as per ecomm spec
- * If category is clothing it will be set to ["product"]
- * @return Content Type array as defined in:
- * - https://developers.facebook.com/docs/facebook-pixel/reference/#object-properties
- */
-const getContentType = (message, defaultValue, categoryToContent) => {
-  const { integrations } = message;
-  if (
-    integrations &&
-    integrations.FacebookPixel &&
-    isObject(integrations.FacebookPixel) &&
-    integrations.FacebookPixel.contentType
-  ) {
-    return integrations.FacebookPixel.contentType;
-  }
-
-  let { category } = message.properties;
-  if (!category) {
-    const { products } = message.properties;
-    if (products && products.length > 0 && Array.isArray(products)) {
-      if (isObject(products[0])) {
-        category = products[0].category;
-      }
-    }
-  } else {
-    if (categoryToContent === undefined) {
-      categoryToContent = [];
-    }
-    const mapped = categoryToContent;
-    const mappedTo = mapped.reduce((filtered, map) => {
-      if (map.from === category) {
-        filtered = map.to;
-      }
-      return filtered;
-    }, "");
-    if (mappedTo.length) {
-      return mappedTo;
-    }
-  }
-  return defaultValue;
-};
 
 /**
  *
@@ -258,140 +205,6 @@ const handleProduct = (message, categoryToContent, valueFieldIdentifier) => {
   };
 };
 
-/** This function transforms the payloads according to the config settings and adds, removes or hashes pii data.
- Also checks if it is a standard event and sends properties only if it is mentioned in our configs.
- @param message --> the rudder payload
-
- {
-      anonymousId: 'c82cbdff-e5be-4009-ac78-cdeea09ab4b1',
-      destination_props: { Fb: { app_id: 'RudderFbApp' } },
-      context: {
-        device: {
-          id: 'df16bffa-5c3d-4fbb-9bce-3bab098129a7R',
-          manufacturer: 'Xiaomi',
-          model: 'Redmi 6',
-          name: 'xiaomi'
-        },
-        network: { carrier: 'Banglalink' },
-        os: { name: 'android', version: '8.1.0' },
-        screen: { height: '100', density: 50 },
-        traits: {
-          email: 'abc@gmail.com',
-          anonymousId: 'c82cbdff-e5be-4009-ac78-cdeea09ab4b1'
-        }
-      },
-      event: 'spin_result',
-      integrations: {
-        All: true,
-        FacebookPixel: {
-          dataProcessingOptions: [Array],
-          fbc: 'fb.1.1554763741205.AbCdEfGhIjKlMnOpQrStUvWxYz1234567890',
-          fbp: 'fb.1.1554763741205.234567890',
-          fb_login_id: 'fb_id',
-          lead_id: 'lead_id'
-        }
-      },
-      message_id: 'a80f82be-9bdc-4a9f-b2a5-15621ee41df8',
-      properties: { revenue: 400, additional_bet_index: 0 },
-      timestamp: '2019-09-01T15:46:51.693229+05:30',
-      type: 'track'
-    }
-
- @param customData --> properties
- { revenue: 400, additional_bet_index: 0 }
-
- @param blacklistPiiProperties -->
- [ { blacklistPiiProperties: 'phone', blacklistPiiHash: true } ] // hashes the phone property
-
- @param whitelistPiiProperties -->
- [ { whitelistPiiProperties: 'email' } ] // sets email
-
- @param isStandard --> is standard if among the ecommerce spec of rudder other wise is not standard for simple track, identify and page calls
- false
-
- @param eventCustomProperties -->
- [ { eventCustomProperties: 'leadId' } ] // leadId if present will be set
-
- */
-
-const transformedPayloadData = (
-  message,
-  customData,
-  blacklistPiiProperties,
-  whitelistPiiProperties,
-  isStandard,
-  eventCustomProperties,
-  integrationsObj
-) => {
-  const defaultPiiProperties = [
-    "email",
-    "firstName",
-    "lastName",
-    "firstname",
-    "lastname",
-    "first_name",
-    "last_name",
-    "gender",
-    "city",
-    "country",
-    "phone",
-    "state",
-    "zip",
-    "birthday"
-  ];
-  blacklistPiiProperties = blacklistPiiProperties || [];
-  whitelistPiiProperties = whitelistPiiProperties || [];
-  eventCustomProperties = eventCustomProperties || [];
-  const customBlackListedPiiProperties = {};
-  const customWhiteListedProperties = {};
-  const customEventProperties = {};
-  for (let i = 0; i < blacklistPiiProperties.length; i += 1) {
-    const singularConfigInstance = blacklistPiiProperties[i];
-    customBlackListedPiiProperties[
-      singularConfigInstance.blacklistPiiProperties
-    ] = singularConfigInstance.blacklistPiiHash;
-  }
-  for (let i = 0; i < whitelistPiiProperties.length; i += 1) {
-    const singularConfigInstance = whitelistPiiProperties[i];
-    customWhiteListedProperties[
-      singularConfigInstance.whitelistPiiProperties
-    ] = true;
-  }
-  for (let i = 0; i < eventCustomProperties.length; i += 1) {
-    const singularConfigInstance = eventCustomProperties[i];
-    customEventProperties[singularConfigInstance.eventCustomProperties] = true;
-  }
-  Object.keys(customData).forEach(eventProp => {
-    const isDefaultPiiProperty = defaultPiiProperties.indexOf(eventProp) >= 0;
-    const isProperyWhiteListed =
-      customWhiteListedProperties[eventProp] || false;
-    if (isDefaultPiiProperty && !isProperyWhiteListed) {
-      delete customData[eventProp];
-    }
-
-    if (
-      Object.prototype.hasOwnProperty.call(
-        customBlackListedPiiProperties,
-        eventProp
-      )
-    ) {
-      if (customBlackListedPiiProperties[eventProp]) {
-        customData[eventProp] =
-          integrationsObj && integrationsObj.hashed
-            ? String(message.properties[eventProp])
-            : sha256(String(message.properties[eventProp]));
-      } else {
-        delete customData[eventProp];
-      }
-    }
-    const isCustomProperty = customEventProperties[eventProp] || false;
-    if (isStandard && !isCustomProperty && !isDefaultPiiProperty) {
-      delete customData[eventProp];
-    }
-  });
-
-  return customData;
-};
 
 const responseBuilderSimple = (
   message,
@@ -429,7 +242,9 @@ const responseBuilderSimple = (
         integrationsObj && integrationsObj.hashed ? split[1] : sha256(split[1]);
     }
     delete userData.name;
+    userData.fbc = userData.fbc ? userData.fbc : deduceFbcParam(message);
   }
+
 
   let customData = {};
   let commonData = {};
